refactor: Navigate to ApprovedDukanScreen after shelf creation
