+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/createShelf/CreateShelfScreen.kt
@@ -28,8 +28,9 @@ import net.thechance.mena.designsystem.presentation.component.text.Text
 import net.thechance.mena.designsystem.presentation.component.textField.TextField
 import net.thechance.mena.designsystem.presentation.theme.theme.MenaTheme
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
+import net.thechance.mena.dukan.presentation.navigation.DukanRoute
 import net.thechance.mena.dukan.presentation.navigation.LocalNavController
-import net.thechance.mena.dukan.presentation.screen.createDukan.content.component.SnackBar
+import net.thechance.mena.dukan.presentation.component.SnackBar
 import net.thechance.mena.dukan.presentation.util.ObserveAsEffect
 import net.thechance.mena.dukan.presentation.util.OnSystemBackPressed
 import net.thechance.mena.dukan.presentation.util.stubPreviews.PreviewCreateShelfInteractionListener
@@ -50,8 +51,9 @@ fun CreateShelfScreen(
     val navController = LocalNavController.current
 
     ObserveAsEffect(viewModel.effect) { effect ->
-        if (effect == CreateShelfEffect.NavigateBack) {
-            navController.popBackStack()
+        when (effect) {
+            CreateShelfEffect.NavigateBack -> navController.popBackStack()
+            CreateShelfEffect.NavigateToApprovedDukan -> navController.navigate(DukanRoute.ApprovedDukanScreenRoute)
         }
     }
