+++ dukan_presentation/src/commonTest/kotlin/net/thechance/mena/dukan/presentation/viewModel/createShelf/CreateShelfViewModelTest.kt
@@ -13,7 +13,7 @@ import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.test.setMain
 import net.thechance.mena.dukan.domain.entity.Shelf
 import net.thechance.mena.dukan.domain.repository.ShelfRepository
-import net.thechance.mena.dukan.presentation.screen.createDukan.content.component.SnackBarType
+import net.thechance.mena.dukan.presentation.component.SnackBarType
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
